opHardeningPluginConfiguration::forceEncodingToUTF8() are now moved as opForceEncodingToUTF8Safeguard
Oh but a standard OpenPNE application implementation is not needed this protection really, because it is blocked by opToolkit::stripNullByteDeep().
opToolkit::stripNullByteDeep() calls preg_replace() with "u" modifier, this denys invalid UTF-8 input ... :(

An advantage of this commit is that this *filters* invalid UTf-8 characters, not *blocks*. Filtering invalid characters are proceeded before calling opToolkit::stripNullByteDeep().